+++ fastai/tabular/data.py
@@ -84,15 +84,18 @@ class TabularDataBunch(DataBunch):
 
     @classmethod
     def from_df(cls, path, df:DataFrame, dep_var:str, valid_idx:Collection[int], procs:OptTabTfms=None,
-                cat_names:OptStrList=None, cont_names:OptStrList=None, classes:Collection=None, **kwargs)->DataBunch:
+                cat_names:OptStrList=None, cont_names:OptStrList=None, classes:Collection=None, 
+                test_df=None, **kwargs)->DataBunch:
         "Create a `DataBunch` from `df` and `valid_idx` with `dep_var`."
         cat_names = ifnone(cat_names, [])
         cont_names = ifnone(cont_names, list(set(df)-set(cat_names)-{dep_var}))
         procs = listify(procs)
-        return (TabularList.from_df(df, path=path, cat_names=cat_names, cont_names=cont_names, procs=procs)
+        src = (TabularList.from_df(df, path=path, cat_names=cat_names, cont_names=cont_names, procs=procs)
                            .split_by_idx(valid_idx)
-                           .label_from_df(cols=dep_var, classes=None)
-                           .databunch())
+                           .label_from_df(cols=dep_var, classes=None))
+        if test_df is not None: src.add_test(TabularList.from_df(test_df, cat_names=cat_names, cont_names=cont_names,
+                                                                 processor = src.train.x.processor))
+        return src.databunch(**kwargs)
 
 class TabularList(ItemList):
     "Basic `ItemList` for tabular data."