+++ fastai/text/data.py
@@ -162,6 +162,9 @@ class TextDataBunch(DataBunch):
         "Create a `TextDataBunch` from DataFrames."
         p_kwargs, kwargs = split_kwargs_by_func(kwargs, _get_processor)
         processor = _get_processor(tokenizer=tokenizer, vocab=vocab, **p_kwargs)
+        if classes is None and label_cols:
+            if len(label_cols) == 0:   classes = [0]
+            elif len(label_cols) > 1:  classes = label_cols
         src = ItemLists(path, TextList.from_df(train_df, path, cols=text_cols, processor=processor),
                         TextList.from_df(valid_df, path, cols=text_cols, processor=processor))
         src = src.label_for_lm() if cls==TextLMDataBunch else src.label_from_df(cols=label_cols, classes=classes, sep=label_delim)
@@ -318,7 +321,7 @@ class TextList(ItemList):
 class LMLabel(CategoryList):
     def predict(self, res): return res
     def reconstruct(self,t:Tensor): return 0
-        
+
 class LMTextList(TextList):
     "Special `TextList` for a language model."
     _bunch = TextLMDataBunch